refactor(Select): use `compare` and `valueComparator` functions from `@/shared`
